restructured scala project according with java
