+++ src/main/scala/io/boson/bson/bsonValue/BsValue.scala
@@ -1,4 +1,4 @@
-package io.boson.bsonValue
+package io.boson.bson.bsonValue
 
 /**
   * Created by Tiago Filipe on 08/11/2017.
