+++ src/test/scala/benchmark/PerformanceTest.scala
@@ -1,8 +1,8 @@
 package benchmark
 
 import bsonLib.BsonObject
-import io.boson.bsonValue.{BsNumber, BsSeq}
-import io.boson.nettyboson.Boson
+import io.boson.bson.bsonImpl.Boson
+import io.boson.bson.bsonValue.{BsNumber, BsSeq}
 import io.boson.scalaInterface.ScalaInterface
 import io.netty.util.ResourceLeakDetector
 import io.vertx.core.json.JsonObject
