+++ src/test/scala/io/boson/BuffersTest.scala
@@ -1,11 +1,13 @@
 package io.boson
 
 import java.nio.ByteBuffer
+
 import bsonLib.BsonObject
-import io.boson.nettyboson.Boson
+import io.boson.bson.bsonImpl.Boson
 import org.junit.runner.RunWith
 import org.scalatest._
 import org.scalatest.junit.JUnitRunner
+
 import scala.collection.mutable.ArrayBuffer
 /**
   * Created by Tiago Filipe on 26/09/2017.
